Added configurtable timeout to waitForDev
When the disk is attached, it takes some time for linux udevd to find
out new device and create /dev/disk/by-path/...
We have a timeout and if it expired we stop wating and try operation
anyways, which may fail

This change introduces the configuration option for this timeout. It is
needed on older ESX (e.g. 6.0 U1) which may be slow in letting Linux
kernel know about new hardware. The change also changes default from 10
to 20 seconds - generally it should not matter since we get /dev/ file
created by Linux much faster

To use, the plugin need to be installed with the proper value in
VDVS_ATTACH_WAIT (in GO Duration format), e.g.:
docker plugin install msterin/docker-volume-vsphere:0.14-attach
--grant-all-permissions --alias=vsphere VDVS_ATTACH_WAIT=30s